post slsx migration code cleanup, plus some v2 code cleanup
