+++ apps/logging/serializers.py
@@ -149,8 +149,7 @@ class SLSRequest(Request):
 
 
 class FHIRRequest(Request):
-    def __init__(self, request, api_ver=None):
-        self.api_ver = api_ver
+    def __init__(self, request):
         super().__init__(request)
 
     def includeAddressFields(self):
@@ -189,7 +188,6 @@ class FHIRRequest(Request):
             "type": "fhir_pre_fetch",
             "uuid": self.uuid(),
             "fhir_id": self.fhir_id(),
-            "api_ver": self.api_ver if self.api_ver is not None else 'v1',
             "includeAddressFields": self.includeAddressFields(),
             "user": self.user(),
             "application": self.application(),
@@ -199,8 +197,7 @@ class FHIRRequest(Request):
 
 
 class FHIRRequestForAuth(Request):
-    def __init__(self, request, auth_flow_dict=None, api_ver=None):
-        self.api_ver = api_ver
+    def __init__(self, request, auth_flow_dict=None):
         if auth_flow_dict:
             self.auth_flow_dict = auth_flow_dict
         else:
@@ -220,7 +217,6 @@ class FHIRRequestForAuth(Request):
         result = {
             "type": "fhir_auth_pre_fetch",
             "uuid": self.uuid(),
-            "api_ver": self.api_ver if self.api_ver is not None else 'v1',
             "includeAddressFields": self.includeAddressFields(),
             "path": "patient search",
             "start_time": self.start_time(),
@@ -268,14 +264,11 @@ class Response:
 class FHIRResponse(Response):
     request_class = FHIRRequest
 
-    def __init__(self, response, api_ver):
-        self.api_ver = api_ver
+    def __init__(self, response):
         super().__init__(response)
 
     def to_dict(self):
         super_dict = super().to_dict()
-        # add fhir version info
-        super_dict.update({"api_ver": self.api_ver if self.api_ver is not None else 'v1'})
         # over write type
         super_dict.update({"type": "fhir_post_fetch"})
         return super_dict
@@ -284,8 +277,7 @@ class FHIRResponse(Response):
 class FHIRResponseForAuth(Response):
     request_class = FHIRRequestForAuth
 
-    def __init__(self, response, auth_flow_dict=None, api_ver=None):
-        self.api_ver = api_ver
+    def __init__(self, response, auth_flow_dict=None):
         if auth_flow_dict:
             self.auth_flow_dict = auth_flow_dict
         else:
@@ -294,8 +286,6 @@ class FHIRResponseForAuth(Response):
 
     def to_dict(self):
         super_dict = super().to_dict()
-        # add fhir version info
-        super_dict.update({"api_ver": self.api_ver if self.api_ver is not None else 'v1'})
         # over write type
         super_dict.update({"type": "fhir_auth_post_fetch"})
         # Update with auth flow session info
