+++ apps/mymedicare_cb/views.py
@@ -1,9 +1,7 @@
-import datetime
 import logging
 import random
 import requests
 import urllib.request as urllib_request
-import waffle
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
@@ -22,13 +20,11 @@ from apps.dot_ext.loggers import (clear_session_auth_flow_trace,
                                   update_instance_auth_flow_trace_with_state)
 from apps.dot_ext.models import Approval
 from apps.fhir.bluebutton.models import hash_hicn, hash_mbi
-from apps.logging.serializers import SLSUserInfoResponse
 from apps.mymedicare_cb.models import (BBMyMedicareCallbackCrosswalkCreateException,
                                        BBMyMedicareCallbackCrosswalkUpdateException)
-from .authorization import OAuth2Config, OAuth2ConfigSLSx
+from .authorization import OAuth2ConfigSLSx
 from .loggers import log_authenticate_start, log_authenticate_success
 from .models import AnonUserState, get_and_update_user
-from .signals import response_hook_wrapper
 from .validators import is_mbi_format_valid, is_mbi_format_synthetic
 
 logger = logging.getLogger('hhs_server.%s' % __name__)
@@ -52,10 +48,7 @@ class BBSLSxHealthCheckFailedException(APIException):
 # For SLS auth workflow info, see apps/mymedicare_db/README.md
 def authenticate(request):
     # Update authorization flow from previously stored state in AuthFlowUuid instance in mymedicare_login().
-    if waffle.switch_is_active('slsx-enable'):
-        request_state = request.GET.get('relay')
-    else:
-        request_state = request.GET.get('state')
+    request_state = request.GET.get('relay')
 
     clear_session_auth_flow_trace(request)
     update_session_auth_flow_trace_from_state(request, request_state)
@@ -64,90 +57,36 @@ def authenticate(request):
     auth_flow_dict = get_session_auth_flow_trace(request)
 
     # SLS vs. SLSx flow based on feature switch slsx-enable (true = SLSx / false = SLS)
-    if waffle.switch_is_active('slsx-enable'):
-        request_token = request.GET.get('req_token', None)
-        if request_token is None:
-            log_authenticate_start(auth_flow_dict, "FAIL",
-                                   "SLSx request_token is missing in callback error.")
-            raise ValidationError(settings.MEDICARE_ERROR_MSG)
-        slsx_client = OAuth2ConfigSLSx()
-        try:
-            access_token, user_id = slsx_client.exchange_for_access_token(request_token, request)
-        except requests.exceptions.HTTPError as e:
-            log_authenticate_start(auth_flow_dict, "FAIL",
-                                   "Token request response error {reason}".format(reason=e))
-            raise BBMyMedicareCallbackAuthenticateSlsClientException(settings.MEDICARE_ERROR_MSG)
+    # removed checking of switch 'slsx-enable' per BB2-397
+    request_token = request.GET.get('req_token', None)
+    if request_token is None:
+        log_authenticate_start(auth_flow_dict, "FAIL",
+                               "SLSx request_token is missing in callback error.")
+        raise ValidationError(settings.MEDICARE_ERROR_MSG)
+    slsx_client = OAuth2ConfigSLSx()
+    try:
+        access_token, user_id = slsx_client.exchange_for_access_token(request_token, request)
+    except requests.exceptions.HTTPError as e:
+        log_authenticate_start(auth_flow_dict, "FAIL",
+                               "Token request response error {reason}".format(reason=e))
+        raise BBMyMedicareCallbackAuthenticateSlsClientException(settings.MEDICARE_ERROR_MSG)
 
-        user_info = slsx_client.get_user_info(access_token, user_id, request)
+    user_info = slsx_client.get_user_info(access_token, user_id, request)
 
-        # Set identity values from userinfo response.
-        sls_subject = user_id.strip()
-        sls_hicn = user_info.get("hicn", "").strip()
-        #     Convert SLS's mbi to UPPER case.
-        sls_mbi = user_info.get("mbi", "").strip().upper()
-        sls_first_name = user_info.get('firstName', "")
-        if sls_first_name is None:
-            sls_first_name = ""
-        sls_last_name = user_info.get('lastName', "")
-        if sls_last_name is None:
-            sls_last_name = ""
-        sls_email = user_info.get('email', "")
-        if sls_email is None:
-            sls_email = ""
-    else:
-        # TODO: Deprecate SLS related code when SLSx is deployed and functioning in all ENVs
-        code = request.GET.get('code')
-        if not code:
-            # Log for info
-            err_msg = "The code parameter is required"
-            log_authenticate_start(auth_flow_dict, "FAIL", err_msg)
-            raise ValidationError(err_msg)
-
-        sls_client = OAuth2Config()
-
-        try:
-            sls_client.exchange(code, request)
-        except requests.exceptions.HTTPError as e:
-            log_authenticate_start(auth_flow_dict, "FAIL",
-                                   "Token request response error {reason}".format(reason=e))
-            raise BBMyMedicareCallbackAuthenticateSlsClientException(settings.MEDICARE_ERROR_MSG)
-
-        userinfo_endpoint = getattr(
-            settings,
-            'SLS_USERINFO_ENDPOINT',
-            'https://test.accounts.cms.gov/v1/oauth/userinfo')
-
-        headers = sls_client.auth_header()
-        # keep using deprecated conv - no conflict issue
-        headers.update({"X-SLS-starttime": str(datetime.datetime.utcnow())})
-        if request is not None:
-            headers.update({"X-Request-ID": str(getattr(request, '_logging_uuid', None)
-                            if hasattr(request, '_logging_uuid') else '')})
-
-        try:
-            response = requests.get(userinfo_endpoint,
-                                    headers=headers,
-                                    verify=sls_client.verify_ssl,
-                                    hooks={
-                                        'response': [
-                                            response_hook_wrapper(sender=SLSUserInfoResponse,
-                                                                  auth_flow_dict=auth_flow_dict)]})
-            response.raise_for_status()
-        except requests.exceptions.HTTPError as e:
-            log_authenticate_start(auth_flow_dict, "FAIL",
-                                   "User info request response error {reason}".format(reason=e))
-            raise BBMyMedicareCallbackAuthenticateSlsClientException(settings.MEDICARE_ERROR_MSG)
-
-        # Get the userinfo response object
-        user_info = response.json()
-
-        # Set identity values from userinfo response.
-        sls_subject = user_info.get("sub", "").strip()
-        sls_hicn = user_info.get("hicn", "").strip()
-        sls_mbi = user_info.get("mbi", "").strip().upper()
-        sls_first_name = user_info.get('given_name', "")
-        sls_last_name = user_info.get('family_name', "")
-        sls_email = user_info.get('email', "")
+    # Set identity values from userinfo response.
+    sls_subject = user_id.strip()
+    sls_hicn = user_info.get("hicn", "").strip()
+    #     Convert SLS's mbi to UPPER case.
+    sls_mbi = user_info.get("mbi", "").strip().upper()
+    sls_first_name = user_info.get('firstName', "")
+    if sls_first_name is None:
+        sls_first_name = ""
+    sls_last_name = user_info.get('lastName', "")
+    if sls_last_name is None:
+        sls_last_name = ""
+    sls_email = user_info.get('email', "")
+    if sls_email is None:
+        sls_email = ""
 
     # If MBI returned from SLS is blank, set to None for hash logging
     if sls_mbi == "":
@@ -206,11 +145,6 @@ def authenticate(request):
 
 
 @never_cache
-def callback_v2(request):
-    return callback(request, 2)
-
-
-@never_cache
 def callback(request, version=1):
     try:
         authenticate(request)
@@ -243,10 +177,8 @@ def callback(request, version=1):
             "error": e.detail,
         }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-    if waffle.switch_is_active('slsx-enable'):
-        state = request.GET.get('relay')
-    else:
-        state = request.GET.get('state')
+    # removed 'slsx-enable' check per BB2-397
+    state = request.GET.get('relay')
 
     if not state:
         return JsonResponse({
@@ -268,9 +200,6 @@ def callback(request, version=1):
     url_map_name = 'oauth2_provider_v2:authorize-instance-v2' if version == 2 else 'oauth2_provider:authorize-instance'
     auth_uri = reverse(url_map_name, args=[approval.uuid])
 
-    # if path.startswith('/v2/o/authorize/') and auth_uri.startswith('/v1/o/authorize/'):
-    #     auth_uri = auth_uri.replace("/v1/o/authorize/", "/v2/o/authorize/")
-
     _, _, auth_path, _, _ = urlsplit(auth_uri)
     return HttpResponseRedirect(urlunsplit((scheme, netloc, auth_path, query_string, fragment)))
 
@@ -282,48 +211,37 @@ def generate_nonce(length=26):
 
 @never_cache
 def mymedicare_login_v2(request):
+    # not invoked by runtime, invoked by unittests only
     return mymedicare_login(request, 2)
 
 
 @never_cache
 def mymedicare_login(request, version=1):
     # SLS vs. SLSx flow based on feature switch slsx-enable (true = SLSx / false = SLS).
-    if waffle.switch_is_active('slsx-enable'):
-        redirect = settings.MEDICARE_SLSX_REDIRECT_URI
-        mymedicare_login_url = settings.MEDICARE_SLSX_LOGIN_URI
+    # removed switch checking 'slsx-enable' per BB2-397, cleanup post SLSX migration complete
+    redirect = settings.MEDICARE_SLSX_REDIRECT_URI
+    mymedicare_login_url = settings.MEDICARE_SLSX_LOGIN_URI
 
-        # Get auth flow session values.
-        auth_flow_dict = get_session_auth_flow_trace(request)
+    # Get auth flow session values.
+    auth_flow_dict = get_session_auth_flow_trace(request)
 
-        # Perform health check on SLSx service
-        slsx_client = OAuth2ConfigSLSx()
-        try:
-            slsx_client.service_health_check()
-        except requests.exceptions.HTTPError as e:
-            log_authenticate_start(auth_flow_dict, "FAIL",
-                                   "SLSx service health check error {reason}".format(reason=e))
-            raise BBSLSxHealthCheckFailedException(settings.MEDICARE_ERROR_MSG)
+    # Perform health check on SLSx service
+    slsx_client = OAuth2ConfigSLSx()
+    try:
+        slsx_client.service_health_check()
+    except requests.exceptions.HTTPError as e:
+        log_authenticate_start(auth_flow_dict, "FAIL",
+                               "SLSx service health check error {reason}".format(reason=e))
+        raise BBSLSxHealthCheckFailedException(settings.MEDICARE_ERROR_MSG)
 
-        relay_param_name = "relay"
-        redirect = urllib_request.pathname2url(redirect)
-        state = generate_nonce()
-        state = urllib_request.pathname2url(state)
-        request.session[relay_param_name] = state
-        mymedicare_login_url = "%s&%s=%s&redirect_uri=%s" % (
-            mymedicare_login_url, relay_param_name, state, redirect)
-        next_uri = request.GET.get('next', "")
-    else:
-        # TODO: Deprecate SLS related code when SLSx is deployed and functioning in all ENVs
-        # check next_uri if is v2 authorize, correct redirect here (do not use default)
-        redirect = settings.MEDICARE_REDIRECT_URI
-        next_uri = request.GET.get('next', "")
-        mymedicare_login_url = settings.MEDICARE_LOGIN_URI
-        redirect = urllib_request.pathname2url(redirect)
-        state = generate_nonce()
-        state = urllib_request.pathname2url(state)
-        request.session['state'] = state
-        mymedicare_login_url = "{}&state={}&redirect_uri={}".format(
-            mymedicare_login_url, state, redirect)
+    relay_param_name = "relay"
+    redirect = urllib_request.pathname2url(redirect)
+    state = generate_nonce()
+    state = urllib_request.pathname2url(state)
+    request.session[relay_param_name] = state
+    mymedicare_login_url = "%s&%s=%s&redirect_uri=%s" % (
+        mymedicare_login_url, relay_param_name, state, redirect)
+    next_uri = request.GET.get('next', "")
 
     AnonUserState.objects.create(state=state, next_uri=next_uri)
