+++ apps/test_urls_reverse.py
@@ -50,9 +50,7 @@ EXPECTED = {
     'my_profile': '/v1/accounts/api/profile',
     'my_profile_v2': '/v2/accounts/api/profile',
     'mymedicare-login': '/mymedicare/login',
-    'mymedicare-login-v2': '/mymedicare/login-v2',
     'mymedicare-sls-callback': '/mymedicare/sls-callback',
-    'mymedicare-sls-callback-v2': '/mymedicare/sls-callback-v2',
     'openid_connect_userinfo': '/v1/connect/userinfo',
     'openid_connect_userinfo_v2': '/v2/connect/userinfo',
     'test_coverage': '/testclient/Coverage',
@@ -111,9 +109,7 @@ URLS_NAMES = {
     "my_profile": None,
     "my_profile_v2": None,
     "mymedicare-login": None,
-    "mymedicare-login-v2": None,
     "mymedicare-sls-callback": None,
-    "mymedicare-sls-callback-v2": None,
     "openid_connect_userinfo": None,
     "openid_connect_userinfo_v2": None,
     "test_coverage": None,
